# on-shop/server/swagger.yml
swagger: "2.0"

info:
  description: "Codecademy e-commerce REST API"
  version: "1.0.0"
  title: "E-commerce REST api"

schemes:
  - http

host: localhost:4000

basePath: "/"

paths:
  /auth/register:
    post:
      summary: "Registers a new user"
      description: "Registers a new user"
      produces:
        - "application/json"
      paramaters:
      responses:
        200:
          description: 200 Success    
          schema:
            type: object
            items:
        409:
          description: 409 Conflict
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
  /auth/login:
    post:
      summary: "Login to a user account"
      description: "Login to a user account"
      produces:
        - "application/json"
      paramaters:
      responses:
        200:
          description: 200 Success    
          schema:
            type: object
            items:
        401:
          description: 401 Unauthorized
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
  /users/{user_id}:
    get:
      summary: "get user record"
      description: "get user record"
      produces:
        - "application/json"
      paramaters:
      responses:
        200:
          description: 200 Success    
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
    put:
      summary: "update user record"
      description: "update user record"
      produces:
        - "application/json"
      paramaters:
      responses:
        200:
          description: 200 Success    
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items: 